Support for SLIDES button.
jekyll-scholar does not have a way of specifying types
of files in the repository with same extensions.  The
format used is the suffix with a "_slides" to indicate
the file used for presentations.  This will work with
PDF files, and PPTX files.

Rendering has to be done on the same line as the PDF
link for the paper.  Have, yet to do that.